+++ src/SME.SR.Application/Queries/RelatorioNotasEConceitosFinais/ObterRelatorioNotasEConceitosFinaisExcel/ObterRelatorioNotasEConceitosFinaisExcelQuery.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using SME.SR.Infra;
+using System.Collections.Generic;
+
+namespace SME.SR.Application
+{
+  public  class ObterRelatorioNotasEConceitosFinaisExcelQuery : IRequest<IEnumerable<RelatorioNotasEConceitosFinaisExcelDto>>
+    {
+        public RelatorioNotasEConceitosFinaisDto RelatorioNotasEConceitosFinais{ get; set; }
+    }
+}
