+++ src/SME.SR.Application/Queries/RelatorioNotasEConceitosFinais/ObterRelatorioNotasEConceitosFinaisPdf/ObterRelatorioNotasEConceitosFinaisPdfQuery.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using SME.SR.Infra;
+
+namespace SME.SR.Application
+{
+    public class ObterRelatorioNotasEConceitosFinaisPdfQuery : IRequest<RelatorioNotasEConceitosFinaisDto>
+    {
+        public FiltroRelatorioNotasEConceitosFinaisDto FiltroRelatorioNotasEConceitosFinais { get; set; }
+
+        public ObterRelatorioNotasEConceitosFinaisPdfQuery(FiltroRelatorioNotasEConceitosFinaisDto filtroRelatorioNotasEConceitosFinais)
+        {
+            this.FiltroRelatorioNotasEConceitosFinais = FiltroRelatorioNotasEConceitosFinais;
+        }
+    }
+}
