+++ src/SME.SR.Infra/Dtos/Relatorios/RelatorioNotasEConceitosFinais/RelatorioNotasEConceitosFinaisDoAlunoDto.cs
@@ -2,16 +2,17 @@
 {
     public class RelatorioNotasEConceitosFinaisDoAlunoDto
     {
-        public RelatorioNotasEConceitosFinaisDoAlunoDto(string turmaNome, string alunoNumeroChamada, string alunoNomeCompleto, string notaConceito)
+        public RelatorioNotasEConceitosFinaisDoAlunoDto(string turmaNome, int alunoCodigo, string alunoNumeroChamada, string alunoNomeCompleto, string notaConceito)
         {
             TurmaNome = turmaNome;
+            AlunoCodigo = alunoCodigo;
             AlunoNumeroChamada = alunoNumeroChamada;
             AlunoNomeCompleto = alunoNomeCompleto;
             NotaConceito = notaConceito;
         }
 
         public string TurmaNome { get; set; }
-         
+        public int AlunoCodigo { get; set; }
         public string AlunoNumeroChamada { get; set; }
         public string AlunoNomeCompleto { get; set; }
         public string NotaConceito { get; set; }
